updated multiplier

+++ causal_effect_estimation/code/BBB_ours.py
@@ -16,7 +16,7 @@ def parallel_helper(n_units):
                          L_TRUE, A_TRUE, Y_TRUE, BURN_IN)
     # print(np.mean(L), np.mean(A), np.mean(Y))
     return estimate_causal_effects_B_B(network_dict, network_adj_mat, L, A, Y, 
-                                       MAX_NEIGHBORS, N_SIM_MULTIPLIER*n_units)
+                                       MAX_NEIGHBORS, int(N_SIM_MULTIPLIER*n_units))
         
 def main():
