+++ causal_effect_estimation/code/setup.py
@@ -100,12 +100,12 @@ def est_w_autog_parallel_helper(params):
     # compute the average causal effect using estimated parameters
     Y_A1_est = estimate_causal_effects_U_U(network_adj_mat, 1, L_est, Y_est, 
                                            burn_in=BURN_IN, 
-                                           n_simulations=N_SIM_MULTIPLIER*n_units,
+                                           n_simulations=int(N_SIM_MULTIPLIER*n_units),
                                            gibbs_select_every=GIBBS_SELECT_EVERY,
                                            L_is_continuous=(L_edge_type == 'B'))
     Y_A0_est = estimate_causal_effects_U_U(network_adj_mat, 0, L_est, Y_est, 
                                            burn_in=BURN_IN, 
-                                           n_simulations=N_SIM_MULTIPLIER*n_units,
+                                           n_simulations=int(N_SIM_MULTIPLIER*n_units),
                                            gibbs_select_every=GIBBS_SELECT_EVERY,
                                            L_is_continuous=(L_edge_type == 'B'))
     return Y_A1_est - Y_A0_est
